requirements.txt
coreapi
coverage==4.5.1
coveralls==1.3.0
dj-database-url==0.5.0
Django==2.0.7
django-import-export==1.0.1
django-nose==1.4.5
django-rest-auth
django-tenant-schemas==1.9.0
djangorestframework==3.8.2
djangorestframework-jwt
gunicorn==19.9.0
Pillow==5.2.0
psycopg2==2.7.5
psycopg2-binary==2.7.5
python-decouple==3.1
whitenoise==3.2
